refact: remove not required route
